Fix menu activation on touchscreens
Removed the click.menubar event binding when autoExpand option is true (which is how Gallery Server uses it). The presence of this binding was causing the menu to close immediately after opening on touchscreens, since touching the menu button triggered these events:  mouseenter, focus, click, and the final click was toggling the menu back to a closed state.